requirements.txt
kafka-python
pyodbc
pandas
streamlit
plotly
streamlit-autorefresh
pytest
flake8
